Convert mask to long tensor in forest dataset return

+++ dataset/forest_dataset.py
@@ -121,4 +121,4 @@ class ForestDataset(Dataset):
             img = transformed["image"]
             mask = transformed["mask"]
 
-        return img, mask
+        return img, mask.long()
